quick fix for (more) userland breakage.
netstat includes a bunch of stuff with _KERNEL defined around it, which
i used here to scope the kernel api. the kernel api needs NULL, which
the kernel gets from sys/param.h, which then pulls in a bunch of machine
stuff and then it all goes horribly wrong on sparc64.

found by deraadt@