Move as much code out of a startup hook as possible
To be able to perform asynchronous operations early we need to
have a robust polling mechanism with flexible condition check
semantics.  A new hv_wait interface provides such facilities.

Right now hvn(4) submits NVS commands and sleeps waiting for a
interrupt to fire and run the completion code and it will take
some effort to convert this code to be able to poll instead of
tsleep.  But VMBus attachment, channel scanning, etc can done
at an earlier stage.  Lets see if this sticks.