Document ERR_load_BN_strings(3). jsing@ confirmed that this function is public and worth documenting.
This page needs much more work, it is outrageously incomplete and
unclear.  For example, it remains unexplained what error strings
are, what "registering" means and what the benefit for the application
is, what happens if it is not done, or what happens if an error
occurs after calling ERR_free_strings(3).  I tried to read the code,
but it is so contorted that i postponed that work.  For example,
it looks like there are hooks for applications to replace the
functions used for registering strings by other, application-supplied
functions, and, of course, there are many levels of macro and
function wrappers.

For now, i only documented the most obvious BUGS.