Improve parsing of the HTTP request line
Make sure that the beginning of a new request starts with an
alphabetic character.  This is a quick way to detect non-ASCII
requests (eg. TLS on port 80).  The full validation of the request
method is done once the input line is read.

Make sure that non-terminated lines do not exceed the
SERVER_MAXHEADERLENGTH which is 8k.  As the current read watermark is
set to 64k, this means that the limit check is triggered after max.
64k of input, depending on the TCP read buffer.

OK benno@ jsing@