deprecate ifq_enqueue_try, and let backends drop arbitrary mbufs.
mikeb@ wants priq to be able to drop lower priority packets if the
current one is high. because ifq avoids freeing an mbuf while an
ifq mutex is held, he needs a way for a backend to return an arbitrary
mbuf to drop rather than signal that the current one needs to be
dropped.

this lets the backends return the mbuf to be dropped, which may or
may not be the current one.

to support this ifq_enqueue_try has to be dropped because it can
only signal about the current mbuf. nothing uses it (except
ifq_enqueue), so we can get rid of it. it wasnt even documented.

this diff includes some tweaks by mikeb@ around the statistics
gathered in ifq_enqueue when an mbuf is dropped.