add ifq_mfreem() so ifq backends can free packets during dequeue.
a goal of the ifq api is to avoid freeing an mbuf while holding a
lock. to acheive this it allowed the backend enqueue operation to
return a single mbuf to be freed. however, mikeb@ is working on a
backend that wants to free packets during dequeue. to support this,
ifq_mfreem queues a packet during dequeue for freeing at the end
of the ifq serialiser.

there's some doco in ifq.h about it.

requested by mikeb@