vmd: Implement simulated baudrate support in the ns8250 module. The previous version was allowing an output rate that is "too fast", and linux guests would give up after 512 characters TXed ("too much work for irq4").
This diff calculates the approximate rate we can sustain at the current
programmed baud rate and limits the output to that rate by inserting a
HZ delay after a specified number of characters have been transmitted.
This fixes the linux guest console issue.

Note that the console now outputs at more or less the selected baud rate,
instead of nearly instantaneously as before - if you selected 9600 in
your guest VMs before, you might want to change that to 115200 now for a
better console experience.

krw@ "seems like a good idea to me"