If we are sending a neighbor solicitation for a link local address send it with a link local source address as well.
This helps upstream routers with their own source address
selection.

A reoccurring scenario is:
- gateway on fe80::1%if
- the gateway does not have an IP in the same prefix as our global address

When we want to talk to the outside world we first need to resolve the
gateway. We copy the source address from our outgoing packet to the
neighbor solicitation packet (a global address) and ask for layer2
information of a link local address.

The upstream router now needs to do source address selection of it's
own. Since we are coming from a global address and there is no address
from the same prefix the router uses another global address lying
around.

We then drop this with "ND packet from non-neighbor".

Reported over the years by a few people, most recently by Marc Peters
on bugs@ who confirmed that this fixes the problem.

OK stsp@, mpi@