In the final RFC 5903 the computation for the DH shared secret changed. Instead of the full point, only the X point is included.
The member g_xy is always the shared secret but so far its buffer has
been allocated using the size of the public points.  Since this is a
different size now, as the shared secret for EC Groups should only store
the x point, we need another member to specify the length of g_xy.

Since this is a backwards incompatible change older isakmpds won't be
able to negotiate if you use EC groups.  Bump the version of our own
vendor tag so peers can try to keep compatibility based on the presen-
ted tag.  This could be used to implement backwards compatibility to
older isakmpds.

Prompted by and ok mpi@